1. TIM5 ADC synchronized with DAC sine wave 2. DAC sine wave activated for testing

+++ DigitalOscilloscope/src/Adc.c
@@ -87,7 +87,7 @@ double Adc_ReadIntTemp() //TODO: synchronize it with other ADC reads
   */
 uint16_t Adc_Read()
 {
-	uint32_t adcData;
+	uint16_t adcData;
 	double voltage;
 
 	//while(!(ADC2->SR & 2)){}
